✨ [feat] 버전 업데이트 판단 로직 추가
