+++ apps/app/app/index.tsx
@@ -52,7 +52,8 @@ const HomeScreen = () => {
         <WebView
           ref={webViewRef}
           source={{
-            uri: process.env.EXPO_PUBLIC_WEB_URL || 'http://127.0.0.1:5173/',
+            // uri: process.env.EXPO_PUBLIC_WEB_URL || 'http://127.0.0.1:5173/',
+            uri: "http://192.168.0.26:5173/"
           }}
           onMessage={bridge.processMessage}
           javaScriptEnabled={true}
