+++ apps/web/src/app/policy/update/useUpdatePolicy.ts
@@ -1,4 +1,7 @@
 import { useEffect } from 'react'
+import { POST_MESSAGE_EVENT } from '@ballog/bridge'
+
+import { useBridge } from '@/shared/hooks/bridge/useBridge'
 
 import { useCheckForUpdate } from './useCheckForUpdate'
 import { useUpdateModal } from './useUpdateModal'
@@ -6,8 +9,12 @@ import { useUpdateModal } from './useUpdateModal'
 // TODO : 버전 api 연동
 export const useUpdatePolicy = () => {
   const { openUpdateModal } = useUpdateModal()
+  const { bridge } = useBridge()
 
-  const { needUpdate } = useCheckForUpdate('1.0.0')
+  const { needUpdate, localVersion } = useCheckForUpdate('1.0.2')
+  bridge.send(POST_MESSAGE_EVENT.SEND_IMAGE_ECHO as any, {
+    웹에서받은version: localVersion,
+  })
 
   useEffect(() => {
     if (!needUpdate) return
@@ -15,5 +22,5 @@ export const useUpdatePolicy = () => {
       openUpdateModal({ type: 'optional' })
     }, 3000)
     return () => clearTimeout(timer)
-  }, [needUpdate, openUpdateModal])
+  }, [needUpdate])
 }
